chore: reorganise jsonrpc modules
